+++ Library/src/sensors/vision/FLS.cpp
@@ -151,7 +151,7 @@ void FLS::NewDataReady(void* data, unsigned int index)
         }
         else
         {
-            sonarData = (GLfloat*)data;
+            sonarData = (GLubyte*)data;
             newDataCallback(this);
             sonarData = NULL;
         }
