+++ Library/src/sensors/vision/SSS.cpp
@@ -152,7 +152,7 @@ void SSS::NewDataReady(void* data, unsigned int index)
         }
         else
         {
-            sonarData = (GLfloat*)data;
+            sonarData = (GLubyte*)data;
             newDataCallback(this);
             sonarData = NULL;
         }
